move sms app utils functions to utils.js
